+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - [TD-6811] Add uid field to LDAP integration
 
+### Fixed
+
+- [TD-6950] Fix add and delete user role functionality
+
 ## [6.12.0] 2024-09-23
 
 ### Added